#11 refactoring crud (#15)
* Create singleton

* Create singleton

* #10Singleton fix

* #11_RefactoryCrud

* fix

* fixed UserService

+++ src/main/java/shop/number/one/model/Order.java
@@ -27,7 +27,7 @@ public class Order {
         addItem(item, 1);
     }
 
-    public void addItem(Item item, int amount) {
+    public void addItem(Item item, long amount) {
         items.put(item, items.getOrDefault(item, 0L) + amount);
     }
 }
